% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workbook.R
\name{clean_workbook}
\alias{clean_workbook}
\title{Erstellt ein neues Sheet in einem Excel file}
\usage{
clean_workbook(workbook_id, workbook_session_id, new_worksheet_name, token)
}
\arguments{
\item{workbook_id}{Die Id der Excel file}

\item{workbook_session_id}{Die Id um Anpassungen in einem Workbook vorzunehmen}

\item{new_worksheet_name}{Der Name des neuen Sheets}

\item{token}{Der Zugriffstoken fuer MS Graph}
}
\value{
Die Funktion gibt die Worksheet_id des neu angelegten Sheets aus
}
\description{
Diese Funktion legt in einem bestehendem Excel file ein neues Sheet an und entfernt alle anderen
}
