% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_folder_from_sharepoint_to_server.R
\name{copy_folder_from_sharepoint_to_server}
\alias{copy_folder_from_sharepoint_to_server}
\title{Diese Funktion kopiert einen Ordner vom Sharepoint in einen Ordner (hier vom Server)}
\usage{
copy_folder_from_sharepoint_to_server(folder_item_id, server_path, token)
}
\arguments{
\item{folder_item_id}{Die Item_Id des zu kopierenden Ordners}

\item{server_path}{Der lokale Pfad an den der Ordner kopiert werden soll}

\item{token}{Der Zugriffstoken für MS_Graph}
}
\value{
Diese Funktion returnt nichts
}
\description{
Diese Funktion kopiert einen Ordner vom Sharepoint in einen Ordner (hier vom Server)
}
