% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_excel_table.R
\name{df_to_excel_table}
\alias{df_to_excel_table}
\title{Dataframe in Excel-Tabelle}
\usage{
df_to_excel_table(
  workbook_session_id,
  workbook_id,
  worksheet_id,
  worksheet_name,
  token,
  dataframe
)
}
\arguments{
\item{workbook_session_id}{Die zuvor erstellte Workbooksession_Id}

\item{workbook_id}{Die Id des Workbooks}

\item{worksheet_id}{Die Id des Sheets in welches der Dataframe geschrieben wird}

\item{worksheet_name}{Der Name des Sheets}

\item{token}{Der Zugriffstoken fuer MS Graph}

\item{dataframe}{Der Dataframe, welcher in das Excel Sheet geschrieben werden soll}
}
\value{
Returnt nichts
}
\description{
Die Funktion schreibt einen Dataframe in ein leeres Excel-Sheet in folgenden Schritten:
\enumerate{
\item Eine Tabelle wird erstellt mit den Zeilen und Spalten des Dataframes, beginnend bei A1, und die erste Zeile
wird der Header mit den Column names des Dataframes
}
}
\details{
\enumerate{
\item Beim Eintragen der Daten wird jede Zeile der Excel Tabelle einzeln addressiert. Die Daten werden Zeile
fuer Zeile eingetragen
}
}
