% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DriveItem_Info.R
\name{get_DriveItem_Info}
\alias{get_DriveItem_Info}
\title{Diese Funktion gibt für eine Datei im Sharepoint Informationen aus}
\usage{
get_DriveItem_Info(absolute_path, token, info = "id")
}
\arguments{
\item{absolute_path}{Pfad beginnen mit /General/...}

\item{token}{Zugriffstoken für MS_Graph}

\item{info}{Die Info. Zum Beispiel "name", "webUrl", "parentReference", "folder" ... als default "id"}
}
\value{
Gibt die gewählte Info zurück
}
\description{
Diese Funktion gibt für eine Datei im Sharepoint Informationen aus
}
