% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DriveItem_Info_delegated.R
\name{get_DriveItem_Info_delegated}
\alias{get_DriveItem_Info_delegated}
\title{DriveItem-Info ueber den delegierten SharePoint-Zugriff}
\usage{
get_DriveItem_Info_delegated(absolute_path, auth, info = "id")
}
\arguments{
\item{absolute_path}{Pfad beginnend mit /General/... (Backslashes vorher zu / aendern).}

\item{auth}{Liste aus authentication_process()$msgraph_sharepoint.}

\item{info}{Die Info, z. B. "name", "webUrl", "parentReference" - Default "id".}
}
\value{
Der gewaehlte Info-Wert des DriveItems.
}
\description{
Delegiertes Pendant zu get_DriveItem_Info(): loest einen absoluten Pfad
(beginnend mit /General/...) im Drive der konfigurierten Site auf.
Wirft bei Nicht-Fund einen Fehler (FlowForce soll Fehlversuche flaggen).
}
