% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folder_structure.R
\name{get_folder_structure}
\alias{get_folder_structure}
\title{Listet alle Elemente eines Ordners als Dataframe auf}
\usage{
get_folder_structure(group_id, folder_id, dataframe, token)
}
\arguments{
\item{group_id}{Die Group / das Team in dem der Ordner liegt}

\item{folder_id}{Der relevante Ordner}

\item{dataframe}{Format muss: Name, Pfad, Driveitem_id, Typ, Parent_id}

\item{token}{Der Zugriffstoken für MS Graph}
}
\value{
Returnt den Dataframe mit den Infos zu allen Elementen
}
\description{
Listet alle Elemente eines Ordners als Dataframe auf
}
