% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sharepoint_data.R
\name{get_sharepoint_data}
\alias{get_sharepoint_data}
\title{Funktion für den Zugriff des aktuellsten Files aus einem Ordner im Sharepoint}
\usage{
get_sharepoint_data(
  folder_path,
  file_name,
  file_type,
  msgraph_key,
  tmp_folder,
  sheet = 1
)
}
\arguments{
\item{folder_path}{A character with the path in the sharepoint which contains the file.}

\item{file_name}{the name of the file that you want to load.A character with the complete name or start of the sharepoint file name}

\item{file_type}{The type of file to load, is passed to read_most_recent_data.  A character defining the type of the file (xlsx, csv, RDS)}

\item{msgraph_key}{the decrypted msgraph key passed to the function  keys$msgraph\link{1}}
}
\value{
The file that was found in the folder when the name was found - message if the file was not found
}
\description{
this function needs the folder, filename, filetype and msgraph key to load the most recent data.
}
