% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_to_excel.R
\name{numeric_to_excel_column}
\alias{numeric_to_excel_column}
\title{Wandelt eine Dezimalzahl in eine Excel-Zahl um}
\usage{
numeric_to_excel_column(zahl)
}
\arguments{
\item{zahl}{Die umzuwandelnde Dezimalzahl}
}
\value{
Gibt die Excel-Zahl aus
}
\description{
Wandelt eine Dezimalzahl in eine Excel-Zahl um
}
\examples{
numeric_to_excel_column(30)
}
