% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_folder_from_server_to_sharepoint.R
\name{copy_folder_from_server_to_sharepoint}
\alias{copy_folder_from_server_to_sharepoint}
\title{Diese Funktion kopiert einen Ordner (hier von Flowforce) und erstellt ihn in einem Sharepoint-Ordner}
\usage{
copy_folder_from_server_to_sharepoint(server_path, folder_item_id, token)
}
\arguments{
\item{server_path}{Der lokale Pfad zum Ordner}

\item{folder_item_id}{Die Item_Id des Ordners in den der Ordner kopiert werden soll}

\item{token}{Der Zugriffstoken für MS_Graph}
}
\value{
Diese Funktion returnt nichts
}
\description{
Diese Funktion kopiert einen Ordner (hier von Flowforce) und erstellt ihn in einem Sharepoint-Ordner
}
